module5-task1

+++ readme/apps/blog/src/app/comment/comment.service.ts
@@ -5,6 +5,7 @@ import { CommentEntity } from './comment.entity';
 import { CommentRepository } from './comment.repository';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
+import { CommentQuery } from './query/comment.query';
 
 @Injectable()
 export class CommentService {
@@ -23,6 +24,10 @@ export class CommentService {
   }
 
   public async deleteComment(id: number): Promise<void> {
+    const existComment = await this.commentRepository.findById(id);
+    if (!existComment) {
+      throw new Error(`Publication with id ${id} doesn't exist`)
+    }
     this.commentRepository.destroy(id);
   }
 
@@ -30,11 +35,15 @@ export class CommentService {
     return this.commentRepository.findById(id);
   }
 
-  public async getComments(id: number): Promise<Comment[]> {
-    return this.commentRepository.find(id);
+  public async getComments(id: number, query: CommentQuery): Promise<Comment[]> {
+    return this.commentRepository.find(id, query);
   }
 
   public async updateComment(id: number, dto: UpdateCommentDto): Promise<Comment> {
+    const existComment = await this.commentRepository.findById(id);
+    if (!existComment) {
+      throw new Error(`Publication with id ${id} doesn't exist`)
+    }
     return this.commentRepository.update(id, { ...dto, updatedAt: new Date() });
   }
 }
